GUI Recolour
A new GUI python tool to recolour selected region only.
The left and right arrow keys on the keyboard will browse frames one by one. The Undo button undoes all changes to the currently viewed frame only, in case you make a mistake.
To use, run Recolour GUI.py in the same directory as recolour.txt and color.pal.
Customise the values in recolour.txt to what you want: in my example it replaces green with orange. For each row, the index on the left will be replaced with the index on the right.